pyproject.toml
[project]
name = "murmur"
version = "0.1.0"
description = "Global voice-to-text injection for macOS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "rumps>=0.4.0",
    "pynput>=1.7.6",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
]

[project.scripts]
murmur = "murmur.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/murmur"]

[tool.uv]
dev-dependencies = []
